Set LANG on winsrv2012r2-amd64
Fixes Cairo build failures:

    Generating cairo.def
    <stdin>:47:1: error: unterminated comment
    Makefile:3072: recipe for target 'cairo.def' failed

cairo.def is generated by grepping Cairo header files and piping the
result to cpp.  grep >= 2.23 can decide at any point in its execution
that it's really processing a binary file, not a text file, and give up
on printing matched lines.  The Cairo headers include the © character,
and if we don't set LANG, grep interprets this as an encoding error and
gives up.